backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database & ORM (psycopg2-binary requires PostgreSQL to be installed)
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary>=2.9.0
pgvector==0.2.4

# Redis & Caching
redis==5.0.1

# Auth & Security
PyJWT==2.8.1
python-multipart==0.0.6
cryptography==41.0.7
passlib[bcrypt]==1.7.4

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# PDF Parsing & NLP
pdfplumber==0.10.3
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0

# AI/ML & Embeddings
anthropic==0.7.7
sentence-transformers==2.2.2
numpy==1.26.2
scipy==1.11.4

# Data Processing
pandas==2.1.3

# Docker & Containers
docker==7.0.0

# Async & Scheduling
httpx==0.25.2
apscheduler==3.10.4
python-dateutil==2.8.2

# Networking & Graph
networkx==3.2

# Testing & Quality
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging & Monitoring
structlog==23.2.0